refactor: migrate checks to tests
